pyproject.toml
[project]
name = "event-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
]
